requirements.txt
django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
pymysql==1.1.0
python-dotenv==1.0.0
djangorestframework_simplejwt==5.2.2
pyjwt
django-sslserver==0.22
drf-yasg==1.21.7
pyyaml==6.0.1
requests==2.32.3